+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Property `File.Stream`
 - Property `FilePath` from Types `Audio`, `Document`, `PhotoSize`, `Sticker`, `Video`, `VideoNote`, and `Voice`
+- Property `Message.NewChatMember`
 
 ## [14.0.0-beta-342] - 2018-01-03